pyproject.toml
[project]
name = "devbase"
version = "2.2.0"
description = "Docker-based Development Environment Manager"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[tool.uv]
package = false

[tool.uv.sources]
